backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
graphene-django==3.2.3
django-graphql-jwt>=0.4.0
graphene-file-upload==1.3.0
django-cors-headers>=3.13.0
channels==4.2.2
django-channels-graphql-ws==1.0.0rc7
